Move doc images and refactor their links
Some of the images in the docs were using absolute urls which are hard
to get right when using baseurl param. I'm moving them to be closer to
their corresponding doc pages and refactoring them to use relative urls
since this is a better approach for images so closely coupled to their
articles.

(References: #21)